:bug: merge conflicts fixed
