+++ src/main/java/com/nailshop/nailborhood/dto/review/response/ShopReviewListLookupResponseDto.java
@@ -1,5 +1,6 @@
 package com.nailshop.nailborhood.dto.review.response;
 
+import com.nailshop.nailborhood.dto.common.PaginationDto;
 import lombok.Builder;
 import lombok.Getter;
 
@@ -9,9 +10,11 @@ import java.util.List;
 @Builder
 public class ShopReviewListLookupResponseDto {
     private List<ShopReviewLookupResponseDto> shopReviewLookupResponseDto;
+    private PaginationDto paginationDto;
 
 
-    public ShopReviewListLookupResponseDto(List<ShopReviewLookupResponseDto> shopReviewLookupResponseDto) {
+    public ShopReviewListLookupResponseDto(List<ShopReviewLookupResponseDto> shopReviewLookupResponseDto, PaginationDto paginationDto) {
         this.shopReviewLookupResponseDto = shopReviewLookupResponseDto;
+        this.paginationDto = paginationDto;
     }
 }
