+++ src/main/java/com/nailshop/nailborhood/repository/artboard/ArtImgRepository.java
@@ -25,4 +25,12 @@ public interface ArtImgRepository extends JpaRepository<ArtImg, Long> {
             "WHERE ai.artImgId = :artImgId ")
     @Modifying(clearAutomatically = true)
     void deleteByArtImgId(@Param("artImgId") Long artImgId ,@Param("status") Boolean status);
+
+    // 매장에 해당 하는 리뷰 대표 이미지 불러오기
+    @Query("SELECT ai.imgPath " +
+            "FROM ArtImg ai " +
+            "LEFT JOIN ai.artRef a " +
+            "LEFT JOIN a.shop s " +
+            "WHERE a.artRefId = :artRefId AND s.shopId = :shopId AND ai.imgNum = 1")
+    String findArtImgByShopIdAndArtRefId(@Param("shopId") Long shopId, @Param("artRefId") Long artRefId);
 }
